Add files via upload
Completely ignores gravity but works pretty well to have a ball bounce around the screen. Displays x and y positions and velocity components. However, Y velocity is wrong. It consistently displays velocities which are inconsistent with direction of travel. X velocity seems to work. Great.
Gravity as I've tried to implement if before seems to lead to ball sticking to the bottom and not being able to lift off, need to figure out how to fix that.